// src/models/DocUploadData.ts
export class DocUploadData{
    syskey : number;
    accountNo : string;
    branchCode: string;
    branchName: string;
    category: string;
    userId: string;
    uploadedDate: string;
    oldFileName: string;
    newFileName: string;
    filePath: string;
    recordStatus: number;
    updatedUserId: string;

}